refactor(112): Rename InnerNavigationBar to DestinationPagination
