+++ src/components/DestinationPagination/animations/horizontal-move.ts
